Move PyQIR test to separate test case
