+++ pip/tests-qir/test_requirements.txt
@@ -1,1 +1,2 @@
 pytest
+pyqir
